GNG-568: Move PageTree into a generic component for re-use
